--- .dockerignore
@@ -1,9 +1,0 @@
-.gitignore
-__pycache__/
-*.pyc
-*.pyo
-*.pyd
-env/
-venv/
-.venv/
-.dockerignore
