+++ .gitignore
@@ -1,3 +1,1 @@
-model/data/
 __pycache__/
-.ipynb_checkpoints/
