--- model/new_model/resnet50_custom_model.py
@@ -1,187 +1,0 @@
-from model.abstract.abstract_model import AbstractModel
-import torch
-import torchvision
-from torchinfo import summary
-from torch import nn
-from pathlib import Path
-from model.utils.data_handler import DataHandler
-from model.utils.saliency_utils import SaliencyGenerator
-
-HARDCODED_CLASS_NAMES = sorted([
-    "airplane", "airport", "baseball_diamond", "basketball_court", "beach",
-    "bridge", "chaparral", "church", "circular_farmland", "cloud",
-    "commercial_area", "dense_residential", "desert", "forest", "freeway",
-    "golf_course", "ground_track_field", "harbor", "industrial_area",
-    "intersection", "island", "lake", "meadow", "medium_residential",
-    "mobile_home_park", "mountain", "overpass", "palace", "parking_lot",
-    "railway", "railway_station", "rectangular_farmland", "river",
-    "roundabout", "runway", "sea_ice", "ship", "snowberg",
-    "sparse_residential", "stadium", "storage_tank", "tennis_court",
-    "terrace", "thermal_power_station", "wetland"
-])
-
-class ResNet50CustomModel(AbstractModel):
-    """This class allows user to create new model basing on the 
-        ResNet-50 architecture. User is able to select what layers to unfreeze and select the 
-        rate of dropout.
-
-    Args:
-        AbstractModel (_type_): abstract model.
-    """
-    def __init__(self,
-                 image_path: str,
-                 dropout_rate: float = 0.5,
-                 unfreeze_specific_blocks: list[str] = None,
-                 unfreeze_classifier: bool = True,
-                 train: str = "train/",
-                 val: str = "val/",
-                 test: str = "test/"
-                 ):
-        """This is the ResNet50CustomModel class constructor.
-
-        Args:
-            dropout_rate (float, optional): dropout rate. Defaults to 0.5.
-            unfreeze_specific_blocks (list[str], optional): layers that need to be unfreezed. Defaults to None.
-            unfreeze_classifier (bool, optional): whether unfreeze classifier or not. Defaults to True.
-        """
-        super().__init__()
-        
-        self.__image_path = Path(image_path)
-        self.__train_dir = self.__image_path / train
-        self.__val_dir = self.__image_path / val
-        self.__test_dir = self.__image_path / test
-
-        _, _, self.__test_dataloader = self.__initializeDataHandler()
-        
-        self.__class_names = HARDCODED_CLASS_NAMES
-
-        self.__device = "cuda" if torch.cuda.is_available() else "cpu"
-        self.__unfreeze_classifier = unfreeze_classifier
-        self.__unfreeze_specific_blocks = unfreeze_specific_blocks
-        self.__dropout_rate = dropout_rate
-
-        weights_enum = torchvision.models.ResNet50_Weights.IMAGENET1K_V2
-        self.__weights_transformed = weights_enum.transforms()
-        self.__model = torchvision.models.resnet50(
-            weights=weights_enum
-            ).to(self.__device)
-        
-        self.__freeze_layers()
-        self.__modify_last_layer()
-        self.__saliency_generator = SaliencyGenerator(self.__model)
-
-    def __initializeDataHandler(self):
-        """This function creates DataHandler class instance."""
-        datahandler = DataHandler(
-            train_dir = str(self.__train_dir), 
-            val_dir=str(self.__val_dir),
-            test_dir = str(self.__test_dir)
-            )
-        train = datahandler.train_dataloader
-        val = datahandler.val_dataloader
-        test = datahandler.test_dataloader
-        return train, val, test
-
-    def __freeze_layers(self):
-        """This method freezes the layers, in case user does not want to trani model."""
-        for param in self.__model.parameters():
-            param.requires_grad = False
-
-        if self.__unfreeze_specific_blocks:
-            for block_name in self.__unfreeze_specific_blocks:
-                if hasattr(self.__model, block_name):
-                    print(f"Unfreezing parameters for: {block_name}")
-                    for param in getattr(self.__model, block_name).parameters():
-                        param.requires_grad = True
-                else:
-                    print(f"Warning: Block '{block_name}' not found in model. Available top-level blocks: "
-                          f"{[name for name, _ in self.__model.named_children()]}")
-
-    def __modify_last_layer(self):
-        """This method modifies last layers of ResNet-50 architecture."""
-        num_ftrs = self.__model.fc.in_features
-        
-        new_classifier = nn.Sequential(
-            nn.Dropout(p=self.__dropout_rate),
-            nn.Linear(num_ftrs, 45)
-        ).to(self.__device)
-
-        self.__model.fc = new_classifier
-
-        if self.__unfreeze_classifier:
-            print("Ensuring the new classifier head is trainable.")
-            for param in self.__model.fc.parameters():
-                param.requires_grad = True
-        else:
-            print("New classifier head will be frozen (parameters will not be updated during training).")
-            for param in self.__model.fc.parameters():
-                param.requires_grad = False
-
-    def forward(self, x):
-        """This is the forward pass of the model."""
-        return self.__model(x)
-    
-    @property
-    def model_summary(self):
-        """This method provides the summary of the model."""
-        return summary(model=self.__model, 
-                input_size=(32, 3, 224, 224),
-                # col_names=["input_size"],
-                col_names=["input_size", "output_size", "num_params", "trainable"],
-                col_width=20,
-                row_settings=["var_names"]
-        )
-    
-    @property
-    def weights(self):
-        """This is a getter for weights."""
-        return self.__weights_transformed
-    
-    @property
-    def model(self):
-        """This is a getter for the model."""
-        return self.__model
-
-    @property
-    def test_accuracy(self):
-        """Evaluates the model on the given test dataset and returns
-          accuracy.
-        """
-        self.__model.eval()
-        correct = 0
-        total = 0
-
-        with torch.no_grad():
-            for images, labels in self.__test_dataloader:
-                images = images.to(self.__device)
-                labels = labels.to(self.__device)
-
-                outputs = self.__model(images)
-                _, predicted = torch.max(outputs.data, 1)
-                total += labels.size(0)
-                correct += (predicted == labels).sum().item()
-
-        test_accuracy = 100 * correct / total
-        return test_accuracy
-        
-    @property
-    def class_names(self):
-        return self.__class_names
-    
-    @property
-    def image_path(self):
-        return self.__image_path
-    
-    @property
-    def test_directory(self):
-        return self.__test_dir
-    
-    def get_saliency_visualization(self, input_tensor: torch.Tensor, target_class_idx: int = None):
-        input_tensor_clone_for_saliency = input_tensor.clone().detach().requires_grad_(True)
-        saliency_map_data, _ = self.__saliency_generator.compute_saliency_map_data(
-            input_tensor_clone_for_saliency, target_class_idx
-        )
-        saliency_viz_b64 = self.__saliency_generator.create_saliency_visualization_base64(
-            input_tensor.clone().detach(), saliency_map_data
-        )
-        return saliency_viz_b64
